First changes toward implementing direct DDE subtract
To be able to subtract the model data, it needs to be stored before the weights are applied. This commit moves weighting toward the MultiDirSolver, such that the DDECal class maintains an unmodified version of the model&data. This does require twice the amount of memory, but I think that's not an issue.