Reuse threads in GainCal's parallel for (#107)
* Reuse threads in GainCal's parallel for

I noticed that GainCal is slower than it used to be. I'm not sure why, but a lot of threads are restarted
during the parallel for structures. This commit keeps the threads running between iterations.

This required a barrier. Unfortunately, the barrier in boost is buggy and the barrier in std requires
c++20 which we probably want to avoid, so I implemented a very simple one myself.

* Don't recreate threadpools during prediction steps

* Fix leaky barrier

* Make beam calculations multi-threaded

...and some other multi-threading fixes.

* Fix wrong function call

* Always turn -O3 and -march=native on

On my testcluster (Dawn), it is not possible to compile in Release mode. After a lot of performance checks
I noticed that this implies everything is compiled with debug settings, which turns out to be a factor
of ~6 slower in my case.