Unsignedint (#141)
* Replace uint by unsigned int

On MacOS Mojave 10.14, there is no standard typedef of unsigned int to
uint. Figuring out how or where to get the typedef turned out to be
harder than just using 'unsigned int' everywhere. I think having a
DPPP-specific types.h just for this is not worth it.

* Disable PythonDPPP and TestDynDPPP on MacOS

Dynamic linking currently does not compile on MacOS, so disable for now.

* Ignore Visual Studio Code files