Make use of itsStepInSolInt consistent
During 'calibrate()', this variable now consistently means 'up to but excluding'.
Unlike DDECal, itsStepInSolInt is not used in calibrate(), so the fact that it
had an inconsistent meaning is not a bug. It seems the solver works fine during
an 'incomplete' finish(), because the data matrices that belong to timesteps
that have not been read are set to zero. Related to #168.