Fix dde subtract one-off bug (#169)
* Fix one-off error when subtract=false in DDECal

When calling 'doSolve()', itsStepInSolInt meant 'up to and including' during regular
process()ing, whereas in finish() it meant 'up to but excluding'. Fixes #168.

* Make use of itsStepInSolInt consistent

During 'calibrate()', this variable now consistently means 'up to but excluding'.
Unlike DDECal, itsStepInSolInt is not used in calibrate(), so the fact that it
had an inconsistent meaning is not a bug. It seems the solver works fine during
an 'incomplete' finish(), because the data matrices that belong to timesteps
that have not been read are set to zero. Related to #168.