Simplify locking for HDF5 (#173)
There was a case where the HDF5 mutex was not set when calling ApplyCal
from nested threads. Rather than trying to pass the mutex around, I've
introduced a static mutex, which also cleans up a bit. This mutex is
used only for reading HDF5, so is not in any critical part of execution.

Solves #167.