Add step for setting beam keywords
This is useful after a predict (with DP3 or image based with e.g. wsclean) in which no consideration of the beam was applied.
In that case, the visibilities are "corrected" for
the beam towards an arbitrary direction. If one source is predicted, it is as if the beam is corrected for in that direction.
This step makes that possible, and e.g. to unapply/corrupt for the beam after having predicted/fted the visibilities afterwards.

This fixes #236.