build: fixup python search and support BLAS_LIBS, LAPACK_LIBS
This is a set of changes we carry at Universität Hamburg to get
the LOFAR toolchain built. I hope it can be included upstream.

Important points are:

-  Use current (less deprecated) cmake ways to detect python.
-  Less redundancy in Python checking in CMakeLists at deeper level
   (Always try to build PythonDPPP as Python and Boost is required
   anyway.)
-  Add support for BLAS_LIBS and LAPACK_LIBS, as this is
   customary in HPC installations to choose the (optimized) BLAS
    implementation in an environment that offers multiple options.

The last point is repeated in other patches we carry for the LOFAR
stack. Maybe the common source for the FindCasacore code should
be patched instead?