Convert FacetPredict to DPStep
Start converting FacetPredict to DPStep

This is still a work in progres..
The step is created but it is not working properly with the directions.
FacetPredict is still determening directions which causes a mismatch in DDECal, hence the failing test.

Continue work on IDGStep

Add documentation

WIP rename FacetPredict to IDGPrect; add integration test