Improve memory of DDECal by avoiding extra modeldata copy
In the Rapthor pipeline, the DDECal solver is often used without subtraction.
In that case, no copy of the model data is required. For large nr of
directions, the model data can be 10s of GB (and is nr of directions
times larger than the observation data). This commit avoids this
extra copy when it is not necessary.

See https://jira.skatelescope.org/browse/RAP-51 .