Initial version of Debian packaging for DP3
This commit contains the initial changes for creating a Debian package
for DP3. The CMake module PythonInstall is no longer used, so it was
removed. The new directory CPack contains the files that make packaging
possible. The changes in CMakeLists.txt make packaging available and
sets the correct path for packaging of Python files. DPPP/CMakeLists.txt
fixes the actual installation problem with Python files.