Enable target CPU selection
Provide more fine-grained control over the selection of the target CPU.
The default behaviour still is to use `-march=native`, but this can be
overridden by the user by setting `TARGET_CPU=<cpu-type>`. The script
DetermineTargetCPU.cmake will try to determine the CPU type that is
identified by the compiler. The identified CPU type will be added to
the package name when building a Debian package.