Fix continuing with non-existing sourcedb in DDECal
When specifying a non-existing sourcedb file in DDECal, an empty sourcedb
is created and no error is reported at that point. Later, DDECal complaints
about having zero directions to solve, which is a bit confusing. This MR
checks whether the sourcedb exists (as was done in OnePredict) and reports
an error if it does not.