Improve the timer class
Use `std::chrono::steady_clock` instead of evaluating the CPU timers.
This leads to a more accurate clock and portable code.

It would be possible to change returned durations to use the clock's duration
type. However that would require changing all callers, instead it still
returns a `double` to be a drop in replacement.

Some minor janitorial changes:
- Make all members variables private
- Use consistent naming for member variables
- Remove the declaration of print_time since it's not defined
- Replace `unsigned long long` with `uint64_t`

The timer tests are moved to timerstest. These tests require multiple
processors else their result will be flaky. Since most of the unit tests
can be exectured in parallel this change only claims the additional
cores when they are needed.